Eliminate unwanted solutions
It prevents the allocation of single links and routers (e.g. from a router to a processor and the corresponding router) without further connection or usage. It has an influence. especially, on the arbitrary hop routing encoding. In the hop bound routing encoding these cases are avoided by the limited hop number. But for completeness, I added it here as well.